pyproject.toml
[project]
name = "ikaros"
version = "0.2.1"
description = "Ikaros is a multi-platform AI assistant with an extension runtime."
readme = "README.md"
requires-python = ">=3.14"
dependencies = []

[dependency-groups]
shared-runtime = [
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0.0,<2.1.0",
]
bot-runtime = [
    { include-group = "shared-runtime" },
    "yt-dlp>=2024.0.0",
    "openai>=2.0.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
    "pyotp>=2.9.0",
    "ddgs>=9.10.0",
    "exa-py>=2.4.0",
]
manager-runtime = [
    { include-group = "bot-runtime" },
    "python-telegram-bot[job-queue]>=22.6",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "python-dateutil>=2.8.0",
    "aiofiles>=23.0.0",
    "mistune>=3.0.0",
    "qrcode>=8.2",
    "cryptography>=46.0.3",
    "discord-py>=2.6.4",
    "dingtalk-stream>=0.24.3",
]
# Heavy or currently-unused skill packs stay opt-in so the default
# manager image remains lean. Install `optional-skill-runtime`
# when you need a full capability image.
skill-web-optional = [
    "googlenewsdecoder>=0.1.5",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.11.0",
]
skill-document-optional = [
    "pypdf>=4.0.0",
    "reportlab>=4.0.0",
    "pdfplumber>=0.10.0",
]
skill-sandbox-optional = [
    "RestrictedPython>=7.0",
]
skill-analysis-optional = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.8.0",
    "yfinance>=0.2.36",
    "scikit-learn>=1.4.0",
    "seaborn>=0.13.0",
    "openpyxl>=3.1.0",
]
skill-integration-optional = [
    "mcp>=1.0.0",
]
memory-provider-optional = [
    "mem0ai>=1.0.1",
    "pymilvus>=2.6.0",
]
optional-skill-runtime = [
    { include-group = "skill-web-optional" },
    { include-group = "skill-document-optional" },
    { include-group = "skill-sandbox-optional" },
    { include-group = "skill-analysis-optional" },
    { include-group = "skill-integration-optional" },
]
manager = [
    { include-group = "manager-runtime" },
]
api = [
    # API 会直接复用 AiService / core.config 做图片记账等能力，
    # 需要 bot-runtime 中的 OpenAI 客户端与相关基础依赖。
    { include-group = "bot-runtime" },
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0,<0.31.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "fastapi-users[oauth,sqlalchemy]>=15.0.3",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20,<0.1.0",
    "toml>=0.10.2",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
default-groups = [
    "dev",
    "manager",
    "api",
]
